enable weak convergence on CI
It's costly, but now with the latest work we will definitely want to be tracking weak convergence orders, so let's get this going. 

I split the tests and did some renaming. Hopefully they all fit in the 50 minutes for one Travis job, otherwise we can split them into separate Travis jobs. If necessary, we can make this use the GPU CI.