enable weak convergence on CI (#271)
* enable weak convergence on CI

It's costly, but now with the latest work we will definitely want to be tracking weak convergence orders, so let's get this going. 

I split the tests and did some renaming. Hopefully they all fit in the 50 minutes for one Travis job, otherwise we can split them into separate Travis jobs. If necessary, we can make this use the GPU CI.

* don't need distributed

* no `@everywhere`

* added using random (#272)

* added using random

* solves "syntax: invalid syntax "1./"; add space(s) to clarify" error"

* get tests running?

* try diffeqfast cpu CI

* Update .gitlab-ci.yml

* update for new SRKs

* fix using

* stricter problem choices

* fix test name

* disable the really long ones

Co-authored-by: frankschae <42201748+frankschae@users.noreply.github.com>